SocialScout: Refactoring and tokenization checks
This PR as facilitation for checks if the message is correctly tokenized. If not, an exception is generated and a message is sent to Twitter user. This reduces traffic to LinkGenerator.
