+++ social-scout-twitter/src/main/java/echoic/socialscouttwitter/core/MusicEntity.java
@@ -1,4 +1,4 @@
-package echoic.socialscouttwitter;
+package echoic.socialscouttwitter.core;
 
 import lombok.Data;
