+++ social-scout-twitter/src/main/java/echoic/socialscouttwitter/core/TokenizationException.java
@@ -1,0 +1,9 @@
+package echoic.socialscouttwitter.core;
+
+public class TokenizationException extends Exception
+{
+    public TokenizationException(String message)
+    {
+        super(message);
+    }
+}
